+++ NEWS.md
@@ -1,3 +1,11 @@
+# hint 1.6.2
+
+* Auto advance to next step when run model completes
+
+# hint 1.6.1
+
+* Makes ADR import button red
+
 # hint 1.6.0
 
 * Show spinner when fetching ADR datasets